Update syslog severity level image and enhance Jinja2 template processing
This commit updates the `syslog_severity_level_distribution.png` image to reflect the latest severity level distribution. Additionally, it adds a separator line in the `jinja2_python.py` script to improve the readability of the output when processing templates. The `index` file in the NTC template directory is modified to uncomment the `cisco_ios_show_ip_interface_brief.template` entry, ensuring it is included in the template index. Lastly, comments in the `example.yaml` file are added for clarity on network configuration parameters.